/(un)stab and /(un)mute commands, useable by global ops (can only be set from irc currently) and mute / stab someone forever until removed
Stab prevents a player from drawing anything or changing any settings. It has both server side and client side limits, so even if you edit the client to allow drawing, it will just cause more desyncing
Mute prevents chatting and /me.
Players are notified of mutes / stabs if they quit and rejoin
Errors in the client loop are now displayed, and the client is disconnected