Prevent c1k's mod from deleting us
This is necessary because the way c1k's mod handles a newer version being installed alongside the one that is embedded is incompatible with the advice the error message the server provides upon a protocol mismatch. This is a problem because the error message is very general and is the preferred way to activate a version that is newer than the bundled one, so it should be correct for all mods.

We can either choose to not support mods that don't conform, or conform to these mods ourselves...