Specify a timeout for starttls
This fixes at least one source of zombie clients: client coroutines getting stuck in starttls.

The new first byte policy requires not only a first byte to be sent, but if a TLS connection is being made, also the TLS handshake succeeding and a first byte to be sent in TLS mode, within the timeout.