Fix a hard to reproduce paste preview crash
I don't have anything particularly good to say here. Tl;dr the code allowed nils to be attempted to be put on the socket in place of mouse position components. This is the problem with "delay line" logic like this; when it's buggy, the bugs are hard to reproduce and to describe.