Fix race between calls to client_i:deduplicate_nick_
The issue was that a client would assume that once the other client with the same name dies, the name is not taken anymore, so it's ok to proceed. Sadly, the reality is that it may be in race with another client for the name, which may have already taken it. The solution is to simply keep waiting for the client holding the name to die.

Strictly speaking, this allows a client to get stuck waiting for the name to be free, because there's no guarantee as to what order clients are woken up to other clients releasing names, but this is an issue I'm willing to live with.